ai_service/requirements.txt
fastapi[standard]==0.124.0
uvicorn[standard]==0.38.0

sqlmodel==0.0.27
asyncpg==0.31.0
psycopg[binary]==3.3.2
alembic==1.17.2

pydantic-settings==2.12.0

python-jose[cryptography]==3.5.0

aiokafka==0.12.0
cramjam==2.11.0

httpx==0.28.1
structlog==24.4.0

# Vector store
qdrant-client==1.12.1

# Embedding model (shared with recommendation_service)
sentence-transformers==3.3.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.0+cpu

# Multipart (voice/media uploads)
python-multipart==0.0.20
aiofiles==24.1.0

# Object storage (voice notes)
aiobotocore==2.15.2
